[yamui] Add FB save/restore code for Qualcomm. Contributes to JB#33093
Since framebuffer is cleared on Qualcomm devices when the display
is switched off we need to preserve its content. Added code
to save/restore FB content in yamui-screensaverd before switching
display off.

Accidentally this solution doesn't work on Jolla Tablet because FB
is cleared just right when yamui-screensaverd opens /dev/fb0.
Because of this we have to use ugly "#ifdef __arm__".

The better solution is to unite yamui and yamui-screensaverd into
a single process having FB device open because of sharing FB across
several processes may not work well on all platforms.
This future version should listen some socket/pipe to accept commands
what to draw on the screen.

Bump version to 1.0.6 in the spec.

Signed-off-by: Igor Zhbanov <igor.zhbanov@jolla.com>