[graphics] Show also images that are larger than display
Images that are larger than screen size are completely ignored by
gr_blit(). This combined with display size variance between different
device types can lead to problems that might be difficult to analyze.

Calculate portion of the image that fits on screen and draw it
instead of skipping the image altogether.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>