[graphics] Add retry loop for drm gr_init()
There is already double-open workaround in use for initializing drm
graphics backend. Looks like most of the times it is not needed,
but sometimes immediate retry does not help.

Retry drm initialization upto 5 times with 100ms in between tries.

Refactor code a bit by splitting fbdev and drm specific bits in
separate functions in an attempt to make control flow easier to follow.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>