[yamui] Add skip display cleanup option and drop --debug . JB#60712
The --debug option may be misleading as it would also do the display
cleanup at quitting. The display cleanup is set to be executed always
and thus, a new option to skip the cleanup is added: --skip-cleanup/-c.

On some devices the cleanup is not needed but on some this gives
benefit of progressing the VT status changes appropriately in order to
systemd logind to function correctly in every case.